pyproject.toml
[tool.poetry]
name = "document-rag"
version = "0.1.0"
description = ""
authors = ["Marton Dobos <marton.dobos@otpeco.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
llama-parse = "^0.4.9"
chromadb = "^0.5.5"
llama-index-vector-stores-chroma = "^0.1.10"
llama-index = "^0.10.64"
torch = "2.2.2"
llama-index-embeddings-huggingface = "^0.2.3"
openai = "^1.40.3"
black = "^24.8.0"
fastapi = {extras = ["standard"], version = "^0.112.0"}
uvicorn = "^0.30.6"
python-multipart = "^0.0.9"
minio = "^7.2.7"
s3fs = "^2024.6.1"
fsspec = "^2024.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

